refactor: event class 이름 통일
