+++ src/js/event/ChangeVideoStatusEvent.js
@@ -1,15 +1,15 @@
 import { mainPagePresenter } from '../presenter/MainPagePresenter';
 import { globalStore } from '../store/VideoStore';
 
-export default class DeleteVideo {
+export default class ChangeVideoStatus {
   constructor(data) {
     this.data = data;
   }
 
-  action() {
+  reaction() {
     this.data.target.closest('li').remove();
-    globalStore.removeLocalStorageVideo(this.data.id);
-    globalStore.removeVideo(this.data.id);
+    globalStore.toggleState(this.data.id);
+    globalStore.toggleWatchedToStarge(this.data.id);
     mainPagePresenter.renderNoVideo();
   }
 }
