+++ src/js/event/EventFactory.js
@@ -1,8 +1,8 @@
 import { EVENT_TYPE } from '../constant';
-import { ChangeTab } from './ChangeTab';
-import ChangeVideoStatus from './ChangeVideoStatus';
-import DeleteVideo from './DeleteVideo';
-import LoadInitialVideoList from './LoadInitialVideoList';
+import { ChangeTab } from './ChangeTabEvent';
+import ChangeVideoStatus from './ChangeVideoStatusEvent';
+import DeleteVideo from './DeleteVideoEvent';
+import LoadInitialVideoList from './LoadInitialVideoListEvent';
 import SaveVideoEvent from './SaveVideoEvent';
 import SearchVideoEvent from './SearchVideoEvent';
 
@@ -10,22 +10,22 @@ export default class EventFactory {
   static generate(name, data) {
     switch (name) {
       case EVENT_TYPE.SEARCH_VIDEO:
-        new SearchVideoEvent(data).action();
+        new SearchVideoEvent(data).reaction();
         break;
       case EVENT_TYPE.SAVE_VIDEO:
-        new SaveVideoEvent(data).action();
+        new SaveVideoEvent(data).reaction();
         break;
       case EVENT_TYPE.LOAD_INITIAL_VIDEO:
-        new LoadInitialVideoList().action();
+        new LoadInitialVideoList().reaction();
         break;
       case EVENT_TYPE.CHANGE_VIDEO_STATUS:
-        new ChangeVideoStatus(data).action();
+        new ChangeVideoStatus(data).reaction();
         break;
       case EVENT_TYPE.CHANGE_TAB:
-        new ChangeTab(data).action();
+        new ChangeTab(data).reaction();
         break;
       case EVENT_TYPE.DELETE_VIDEO:
-        new DeleteVideo(data).action();
+        new DeleteVideo(data).reaction();
         break;
       default:
         alert('NO_EVENT');
